refactor: separate executing a lap from main function to enhance code readability
